+++ packages/utils/src/functions/index.ts
@@ -16,3 +16,4 @@ export * from "./nformatter";
 export * from "./time-ago";
 export * from "./truncate";
 export * from "./urls";
+export * from "./tag-colors";
