+++ packages/utils/src/functions/tag-colors.ts
@@ -1,0 +1,33 @@
+import { TagColorProps } from "../types";
+
+export const COLORS_LIST: { color: TagColorProps; css: string }[] = [
+  {
+    color: "red",
+    css: "bg-red-100 text-red-600",
+  },
+  {
+    color: "yellow",
+    css: "bg-yellow-100 text-yellow-600",
+  },
+  {
+    color: "green",
+    css: "bg-green-100 text-green-600",
+  },
+  {
+    color: "blue",
+    css: "bg-blue-100 text-blue-600",
+  },
+  {
+    color: "purple",
+    css: "bg-purple-100 text-purple-600",
+  },
+  {
+    color: "brown",
+    css: "bg-brown-100 text-brown-600",
+  },
+];
+
+export function randomBadgeColor() {
+  const randomIndex = Math.floor(Math.random() * COLORS_LIST.length);
+  return COLORS_LIST[randomIndex].color;
+}
