+++ packages/utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./constants";
 export * from "./functions";
+export * from "./types";
